requirements.txt
wheel==0.36.2
Flask==2.0.1
Flask-Cors==3.0.10
SQLAlchemy==1.4.21
gunicorn==20.1.0
Flask-SQLAlchemy==2.5.1
# when we deploy, this requirement must be uncommented
psycopg2-binary==2.9.3
# psycopg2==2.9.6
werkzeug==2.0.2
Flask-RESTful==0.3.9
python-dotenv==0.20.0
requests==2.27.1
pdfkit==1.0.0
Pillow==9.0.0
webargs==8.2.0
marshmallow==3.17.1
flask-apispec==0.11.4
python-dateutil==2.8.2
apispec==5.2.2
PyJWT==1.7.1
bcrypt==3.2.0
discord.py==1.7.3
PyMySQL==0.9.3
Flask-MySQL==1.4.0
